backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.1
cryptography==41.0.7
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
pyOpenSSL>=23.0.0 